btn like/ delete components
