+++ src/components/BtnDelete/BtnDelete.module.css
@@ -1,0 +1,18 @@
+.delete {
+  position: absolute;
+  z-index: 1;
+  font-size: 18px;
+  right: 20px;
+  top: 20px;
+  color: rgba(255, 255, 255, 0.87);
+  cursor: pointer;
+}
+
+.delete:hover {
+  color: rgb(255, 255, 255);
+  text-shadow: rgb(255, 255, 255) 10px 10px 10px;
+}
+
+.delete:active {
+  color: rgba(0, 0, 0, 0.7);
+}
