+++ src/components/BtnLike/BtnLike.module.css
@@ -1,0 +1,12 @@
+.heart {
+  position: absolute;
+  font-size: 28px;
+  right: 30px;
+  bottom: 30px;
+  color: rgba(255, 255, 255, 0.87);
+  cursor: pointer;
+}
+
+.heart:hover {
+  color: rgba(228, 16, 118, 0.9);
+}
