+++ src/pages/Heroes/Heroes.module.css
@@ -95,16 +95,3 @@
   color: rgba(255, 255, 255, 0.6);
   overflow-wrap: break-word;
 }
-
-.heart {
-  position: absolute;
-  font-size: 28px;
-  right: 30px;
-  bottom: 30px;
-  color: rgba(255, 255, 255, 0.87);
-  cursor: pointer;
-}
-
-.heart:hover {
-  color: rgba(228, 16, 118, 0.9);
-}
